+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for dedicated server
 - Fixed the Black Dick bug
 - Made it possible for people to change what the girls say with the lang file
+- Added the ability to change the color of the penis with the mod menu
 
 ## 0.5.2
 - Fixed a bug where the girls wouldn't spawn in the villages.